Refactor RemoveKeywordCollisions->ManipulateNames
Generalize the operations of the RemoveKeywordCollisions transform
into a new ManipulateNames transform. The ManipulateNames transform is
an abstract transform for making conditional modifications to
keywords/names in a FIRRTL circuit.
